Add blockMap, modified double to float

+++ Platform2D_IGPE/src/element/Position.java
@@ -1,21 +1,27 @@
 package element;
 
 public class Position {
-	private double x,y;
-	public Position(double x,double y) {
+
+	private float x, y;
+
+	public Position(float x, float y) {
 		this.setX(x);
 		this.setY(y);
 	}
-	public double getY() {
+
+	public float getY() {
 		return y;
 	}
-	public void setY(double y) {
+
+	public void setY(float y) {
 		this.y = y;
 	}
-	public double getX() {
+
+	public float getX() {
 		return x;
 	}
-	public void setX(double x) {
+
+	public void setX(float x) {
 		this.x = x;
 	}
 }
